chore: author & sign generated ci commit (#5149)
- Refactor/simplify gh app authentication
- Rename coveo/release private package to coveo/ci to reflect the
increased scope (not only release)
- Add a commit-all script & use it for generated files
