+++ .github/workflows/prbot.yml
@@ -5,19 +5,27 @@ on:
 env:
   CYPRESS_VERIFY_TIMEOUT: 60000
 jobs:
-  report-size:
+  pr-report:
     if: github.event_name == 'pull_request'
-    name: 'Report bundle size'
+    name: 'PR Report'
     runs-on: ubuntu-latest
+    environment: PR Artifacts
     env:
       NODE_OPTIONS: --max_old_space_size=4096
-      GITHUB_CREDENTIALS: ${{ secrets.GITHUB_TOKEN }}
     steps:
       - name: Harden Runner
         uses: step-security/harden-runner@4d991eb9b905ef189e4c376166672c3f2f230481 # v2.11.0
         with:
           egress-policy: audit
-
+      - name: Generate a token
+        if: ${{ always() && github.event_name == 'pull_request'}}
+        id: generate-token
+        uses: actions/create-github-app-token@21cfef2b496dd8ef5b904c159339626a10ad380e # v1
+        with:
+          app-id: ${{ secrets.GH_APP_ID }}
+          private-key: ${{ secrets.GH_APP_PRIVATE_KEY }}
+          owner: coveo
+          repositories: 'ui-kit'
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
         with:
           fetch-depth: 0
@@ -26,9 +34,12 @@ jobs:
           node-version-file: '.nvmrc'
       - run: npm ci
       - run: npm run pr:report
+        env:
+          GITHUB_CREDENTIALS: ${{ steps.generate-token.outputs.token }}
   build:
     name: 'Build'
     runs-on: ubuntu-latest
+    environment: PR Artifacts
     steps:
       - name: Harden Runner
         uses: step-security/harden-runner@4d991eb9b905ef189e4c376166672c3f2f230481 # v2.11.0
@@ -36,8 +47,32 @@ jobs:
           egress-policy: audit
 
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
+        with:
+          ref: ${{ github.head_ref }}
       - uses: ./.github/actions/build
-      - uses: ./.github/actions/commit-generated-files
+      - name: Check for uncommitted changes
+        id: check-changes
+        run: |
+          if [[ -n "$(git status --porcelain)" ]]; then
+            echo "has_changes=true" >> $GITHUB_OUTPUT
+          else
+            echo "has_changes=false" >> $GITHUB_OUTPUT
+          fi
+      - name: Generate a token
+        if: ${{ steps.check-changes.has_changes == 'true' && github.event_name == 'pull_request'}}
+        id: generate-token
+        uses: actions/create-github-app-token@21cfef2b496dd8ef5b904c159339626a10ad380e # v1
+        with:
+          app-id: ${{ secrets.GH_APP_ID }}
+          private-key: ${{ secrets.GH_APP_PRIVATE_KEY }}
+          owner: coveo
+          repositories: 'ui-kit'
+      - name: Commit files
+        if: ${{ steps.check-changes.has_changes == 'true' && github.event_name == 'pull_request'}}
+        run: npm run-script -w=@coveo/ci commit-all
+        env:
+          GITHUB_INSTALLATION_TOKEN: ${{ steps.generate-token.outputs.token }}
+          DEBUG: '*'
   build-cdn:
     name: 'Build CDN'
     runs-on: ubuntu-latest
