+++ packages/atomic-angular/scripts/bump.mjs
@@ -1,4 +1,4 @@
 import {resolve} from 'node:path';
 
 process.env.INIT_CWD = resolve('./projects/atomic-angular');
-await import('@coveo/release/bump-package.mjs');
+await import('@coveo/ci/bump-package.mjs');
