+++ packages/atomic-angular/scripts/publish-npm.mjs
@@ -1,4 +1,4 @@
 import {resolve} from 'node:path';
 
 process.env.INIT_CWD = resolve('./projects/atomic-angular/dist');
-await import('@coveo/release/npm-publish-package.mjs');
+await import('@coveo/ci/npm-publish-package.mjs');
