+++ packages/atomic-hosted-page/package.json
@@ -25,9 +25,9 @@
     "build": "nx build",
     "start": "node ./scripts/start.js",
     "e2e": "playwright test",
-    "publish:npm": "npm run-script -w=@coveo/release npm-publish",
-    "publish:bump": "npm run-script -w=@coveo/release bump",
-    "promote:npm:latest": "npm run-script -w=@coveo/release promote-npm-prod"
+    "publish:npm": "npm run-script -w=@coveo/ci npm-publish",
+    "publish:bump": "npm run-script -w=@coveo/ci bump",
+    "promote:npm:latest": "npm run-script -w=@coveo/ci promote-npm-prod"
   },
   "dependencies": {
     "@coveo/bueno": "1.0.10",
@@ -35,7 +35,7 @@
     "lit": "3.2.1"
   },
   "devDependencies": {
-    "@coveo/release": "1.0.0",
+    "@coveo/ci": "1.0.0",
     "@playwright/test": "1.50.1",
     "@types/node": "22.13.5",
     "rollup": "4.34.8",
