+++ packages/atomic/package.json
@@ -64,9 +64,9 @@
     "e2e:snapshots:watch": "cypress open --config-file cypress-screenshots.config.mjs --browser chrome",
     "e2e:insight": "cypress run --config-file cypress-insight-panel.config.mjs --browser chrome",
     "e2e:insight:watch": "cypress open --config-file cypress-insight-panel.config.mjs --browser chrome",
-    "publish:npm": "npm run-script -w=@coveo/release npm-publish",
-    "publish:bump": "npm run-script -w=@coveo/release bump",
-    "promote:npm:latest": "npm run-script -w=@coveo/release promote-npm-prod",
+    "publish:npm": "npm run-script -w=@coveo/ci npm-publish",
+    "publish:bump": "npm run-script -w=@coveo/ci bump",
+    "promote:npm:latest": "npm run-script -w=@coveo/ci promote-npm-prod",
     "validate:definitions": "tsc --noEmit --esModuleInterop --skipLibCheck ./dist/types/components.d.ts"
   },
   "dependencies": {
@@ -89,7 +89,7 @@
   },
   "devDependencies": {
     "@axe-core/playwright": "4.10.1",
-    "@coveo/release": "1.0.0",
+    "@coveo/ci": "1.0.0",
     "@custom-elements-manifest/analyzer": "0.10.4",
     "@playwright/test": "1.50.1",
     "@rollup/plugin-alias": "5.1.1",
