+++ packages/atomic/project.json
@@ -36,7 +36,7 @@
           "node ./scripts/stencil-proxy.mjs",
           "node ./scripts/build.mjs --config=tsconfig.lit.json",
           "node ./scripts/process-css.mjs --config=tsconfig.lit.json ",
-          "npm run-script -w=@coveo/release if-cdn -- \"rollup -c rollup.config.js\"",
+          "npm run-script -w=@coveo/ci if-cdn -- \"rollup -c rollup.config.js\"",
           "esbuild src/autoloader/index.ts --format=esm --outfile=dist/atomic/autoloader/index.esm.js",
           "esbuild src/autoloader/index.ts --format=cjs --outfile=dist/atomic/autoloader/index.cjs.js",
           "rimraf ./dist/atomic/loader/package.json",
