+++ packages/headless/package.json
@@ -133,8 +133,8 @@
     "test:watch": "vitest --exclude \"src/integration-tests/**\"",
     "integration-test": "vitest run --poolOptions.threads.singleThread src/integration-tests/**",
     "integration-test:watch": "vitest --poolOptions.threads.singleThread src/integration-tests/**",
-    "publish:npm": "npm run-script -w=@coveo/release npm-publish",
-    "publish:bump": "npm run-script -w=@coveo/release bump",
+    "publish:npm": "npm run-script -w=@coveo/ci npm-publish",
+    "publish:bump": "npm run-script -w=@coveo/ci bump",
     "build:typedoc": "npm run build:typedoc:plugin && npm run build:typedoc:docs && npm run build:typedoc:merge",
     "build:typedoc:plugin": "npx tsc -p typedoc/tsconfig.json",
     "build:typedoc:docs": "concurrently \"npm run build:typedoc:search\" \"npm run build:typedoc:commerce\" \"npm run build:typedoc:case-assist\" \"npm run build:typedoc:insight\" \"npm run build:typedoc:recommendation\" \"npm run build:typedoc:ssr\" \"npm run build:typedoc:ssr-commerce\"",
@@ -146,7 +146,7 @@
     "build:typedoc:recommendation": "typedoc --json ./temp/recommendation.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/recommendation.typedoc.json",
     "build:typedoc:ssr": "typedoc --json ./temp/ssr.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/ssr.typedoc.json",
     "build:typedoc:ssr-commerce": "typedoc --json ./temp/ssr-commerce.json --tsconfig tsconfig.typedoc.json --options ./typedoc/configs/ssr-commerce.typedoc.json",
-    "promote:npm:latest": "npm run-script -w=@coveo/release promote-npm-prod"
+    "promote:npm:latest": "npm run-script -w=@coveo/ci promote-npm-prod"
   },
   "peerDependencies": {
     "encoding": "^0.1.13",
@@ -176,7 +176,7 @@
     "ts-debounce": "4.0.0"
   },
   "devDependencies": {
-    "@coveo/release": "1.0.0",
+    "@coveo/ci": "1.0.0",
     "esbuild-plugin-alias-path": "2.0.2",
     "eslint-plugin-canonical": "4.18.0",
     "execa": "9.5.2",
