+++ packages/quantic/package.json
@@ -41,9 +41,9 @@
     "publish:sfdx": "../../node_modules/.bin/ts-node scripts/build/create-package.ts --remove-translations",
     "promote:sfdx": "npm run publish:sfdx -- --promote",
     "promote:sfdx:ci": "npm run publish:sfdx -- --promote --ci",
-    "publish:npm": "npm run-script -w=@coveo/release npm-publish",
-    "publish:bump": "npm run-script -w=@coveo/release bump",
-    "promote:npm:latest": "npm run-script -w=@coveo/release promote-npm-prod",
+    "publish:npm": "npm run-script -w=@coveo/ci npm-publish",
+    "publish:bump": "npm run-script -w=@coveo/ci bump",
+    "promote:npm:latest": "npm run-script -w=@coveo/ci promote-npm-prod",
     "preinstall": "node scripts/npm/check-sfdx-project.js",
     "postinstall": "node scripts/npm/setup-quantic.js",
     "babel:headless": "babel ../../node_modules/@coveo/headless/dist/quantic --out-dir .tmp/quantic-compiled --extensions .js --minified"
@@ -63,7 +63,7 @@
     "@babel/plugin-transform-async-to-generator": "7.25.9",
     "@ckeditor/jsdoc-plugins": "43.0.1",
     "@coveo/relay-event-types": "14.3.3",
-    "@coveo/release": "1.0.0",
+    "@coveo/ci": "1.0.0",
     "@octokit/graphql": "8.2.1",
     "@octokit/graphql-schema": "15.25.0",
     "@playwright/test": "1.50.1",
