+++ packages/samples/headless-react/project.json
@@ -8,7 +8,7 @@
       "cache": true,
       "options": {
         "commands": [
-          "npm run-script -w=@coveo/release if-not-cdn -- \"npm run build:client\""
+          "npm run-script -w=@coveo/ci if-not-cdn -- \"npm run build:client\""
         ],
         "parallel": true,
         "cwd": "{projectRoot}"
