+++ packages/samples/headless-ssr-commerce/project.json
@@ -7,7 +7,7 @@
       "outputs": ["{projectRoot}/.next", "{projectRoot}/next-env.d.ts"],
       "cache": true,
       "options": {
-        "command": "npm run-script -w=@coveo/release if-not-cdn -- \"npm run build:next\"",
+        "command": "npm run-script -w=@coveo/ci if-not-cdn -- \"npm run build:next\"",
         "cwd": "{projectRoot}"
       }
     }
