+++ utils/ci/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@coveo/release",
+  "name": "@coveo/ci",
   "private": true,
   "authors": "Coveo",
   "version": "1.0.0",
@@ -22,13 +22,13 @@
     "typescript": "5.8.2"
   },
   "scripts": {
+    "commit-all": "node ./commit-all.mjs",
     "promote-npm-prod": "node ./promote-npm-tag-to-latest.mjs",
     "git-lock": "node ./git-lock.mjs",
     "bump": "node ./bump-package.mjs",
     "bump:root": "node ./bump-root.mjs",
     "npm-publish": "node ./npm-publish-package.mjs",
     "git-publish-all": "node ./git-publish-all.mjs",
-    "if-cdn": "node ./if-cdn.mjs",
     "if-not-cdn": "node ./if-not-cdn.mjs",
     "if-cdn": "node ./if-cdn.mjs",
     "reify": "node ./reify.mjs",
